% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0docs.R
\name{add_}
\alias{add_}
\title{Add functions}
\arguments{
\item{ont}{An ontology of class \link[simona]{ontology_DAG}.}

\item{terms}{A vector of ontology term IDs.}

\item{g}{\link[tidygraph]{tbl_graph} object.}

\item{force_new}{Add the data again even if the associated column already 
exists.}

\item{lvl}{Depth of the ancestor terms to add. 
Will get the closest ancestor to this level if none have this exact level.}

\item{add_ancestors}{Add ancestors for each term.}

\item{add_n_edges}{Add the number of edges (connections) for each term.}

\item{add_ontology_levels}{Add the ontology level for each term.}
}
\value{
Added data.
}
\description{
Functions to add extra metadata to an ontology or data.table object.
}
\concept{add_}
