% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ancestors.R
\name{add_ancestors}
\alias{add_ancestors}
\title{Add ancestors}
\usage{
add_ancestors(
  ont,
  lvl = 2,
  include_self = TRUE,
  prefix = "ancestor",
  fill_na = TRUE,
  force_new = FALSE
)
}
\arguments{
\item{ont}{An ontology of class \link[simona]{ontology_DAG}.}

\item{lvl}{Depth of the ancestor terms to add. 
Will get the closest ancestor to this level if none have this exact level.}

\item{include_self}{For \code{dag_offspring()} and \code{dag_ancestors()}, this controls whether to also include the query term itself.}

\item{prefix}{A prefix for the ancestor column names.}

\item{fill_na}{If \code{TRUE}, fill ancestor columns with 
NA values using the term ID and name.}

\item{force_new}{Add the data again even if the associated column already 
exists.}
}
\description{
For each term, get its ancestor at a given level
and add the ID and name of the ancestor to the ontology metadata.
By default, lower numbers are higher in the ontology
 (e.g. level 0 includes the most broad ontology term).
When a terms does not have an ancestor at a given level
 (e.g. \code{lvl=3} but the term itself is at level 1) the term is assigned
  to itself as its own ancestor.
When a term has multiple ancestors at a given level, 
the first ancestor is arbitrarily chosen.
}
\examples{
ont <- get_ontology("hp")
ont2 <- add_ancestors(ont, lvl=4)
}
