% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_db.R
\name{add_db}
\alias{add_db}
\title{Add database}
\usage{
add_db(
  dat,
  input_col,
  output_col = paste0(input_col, "_db"),
  split = ":",
  keep = 1
)
}
\arguments{
\item{dat}{data.table or tbl_graph.}

\item{input_col}{Column containing IDs.}

\item{output_col}{Column to add.}

\item{split}{
    character vector (or object which can be coerced to such)
    containing \link[base]{regular expression}(s) (unless \code{fixed = TRUE})
    to use for splitting.  If empty matches occur, in particular if
    \code{split} has length 0, \code{x} is split into single characters.
    If \code{split} has length greater than 1, it is re-cycled along
    \code{x}.
  }

\item{keep}{Specify indices corresponding to just those list elements to retain in the transposed result. Default is to return all.}
}
\description{
Extract a database identifier from an ID column and add it
as a separate column.
}
