% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hoverboxes.R
\name{add_hoverboxes}
\alias{add_hoverboxes}
\title{Add hoverboxes}
\usage{
add_hoverboxes(
  g,
  columns = get_graph_colnames(g),
  hoverbox_column = c("hover", "title", "label"),
  width = 60,
  digits = 3,
  decorators = c("<b>", "</b>"),
  as_html = TRUE,
  force_new = FALSE
)
}
\arguments{
\item{g}{\link[tidygraph]{tbl_graph} object.}

\item{columns}{Character vector of column names to include in the hoverbox.}

\item{hoverbox_column}{Name of the new hoverbox column to add.}

\item{width}{Positive integer giving target line width (in number of
characters). A width less than or equal to 1 will put each word on its
own line.}

\item{digits}{integer indicating the number of decimal places
    (\code{round}) or significant digits (\code{signif}) to be used.
    For \code{round}, negative values are allowed (see \sQuote{Details}).}

\item{decorators}{Left and right decorators to add to the hoverbox titles.}

\item{as_html}{Provide the hoverbox in HTML format.}

\item{force_new}{Add the data again even if the associated column already 
exists.}
}
\value{
tidygraph with additional metadata column named "hover".
}
\description{
A hoverbox is a box of text that shows up when the cursor
 hovers over something.
These can be useful when making interactive network plots
of the HPO phenotypes because we can include a hoverbox that gives
information and data associated with each phenotype.
}
\examples{
ont <- get_ontology("hp")
g <- ontology_to(ont, to="tbl_graph")
g2 <- add_hoverboxes(g)
}
