% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ontology_metadata.R
\name{add_ontology_metadata}
\alias{add_ontology_metadata}
\title{Add ontology metadata}
\usage{
add_ontology_metadata(
  ont,
  lvl = 2,
  force_new = FALSE,
  add_n_edges = TRUE,
  add_ontology_levels = TRUE
)
}
\arguments{
\item{ont}{An ontology of class \link[simona]{ontology_DAG}.}

\item{lvl}{Depth of the ancestor terms to add. 
Will get the closest ancestor to this level if none have this exact level.}

\item{force_new}{Add the data again even if the associated column already 
exists.}

\item{add_n_edges}{Add the number of edges (connections) for each term.}

\item{add_ontology_levels}{Add the ontology level for each term.}
}
\description{
Add per-term metadata to ontology.
}
\examples{
ont <- get_ontology("hp", terms=10)
ont2 <- add_ontology_metadata(ont)
}
