% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_docs.R, R/dt_to_matrix.R, R/graph_to_dt.R,
%   R/graph_to_ggnetwork.R, R/ontology_to.R, R/ontology_to_adjacency.R,
%   R/ontology_to_tidygraph.R, R/to_graph.R
\name{convert_}
\alias{convert_}
\alias{dt_to_matrix}
\alias{graph_to_dt}
\alias{graph_to_ggnetwork}
\alias{ontology_to}
\alias{ontology_to_adjacency}
\alias{ontology_to_graph}
\alias{to_graph}
\title{Converter functions}
\usage{
dt_to_matrix(dat, omit_cols = NULL, rownames = NULL, as_sparse = TRUE)

graph_to_dt(g, what = c("nodes", "edges"), id_col = c("id", "name"))

graph_to_ggnetwork(g)

ontology_to(
  ont,
  to = c("adjacency", "adjacency_dist", "adjacency_dist_hclust",
    "adjacency_dist_hclust_dendrogram", "adjacency_dist_hclust_clusters", "similarity",
    "dendrogram", "igraph", "dot", "adjacency_igraph", "igraph_dist",
    "igraph_dist_hclust", "igraph_dist_hclust_dendrogram", "tbl_graph", "data.frame",
    "data.table", "list"),
  terms = ont@terms,
  remove_terms = grep(":", terms, invert = TRUE, value = TRUE),
  as_sparse = FALSE,
  ...
)

ontology_to_adjacency(
  ont,
  terms = ont@terms,
  remove_terms = grep(":", terms, invert = TRUE, value = TRUE)
)

ontology_to_graph(ont, ...)

to_graph(g, ...)
}
\arguments{
\item{as_sparse}{Return the object as a \link[Matrix]{sparseMatrix}.}

\item{g}{\link[tidygraph]{tbl_graph} object.}

\item{id_col}{Column containing the unique identifier for each node 
in a graph (e.g. "name").}

\item{ont}{An ontology of class \link[simona]{ontology_DAG}.}

\item{to}{A character string specifying the format to convert to.}

\item{terms}{Term IDs to include. Can alternatively be an integer, 
which will be used to randomly sample N terms from the data.}

\item{remove_terms}{Character vector of term IDs to exclude.}

\item{...}{Additional arguments passed to plot-specific functions.}

\item{as}{A character string specifying the format to convert to.}

\item{as_dt}{Return the object as a \link[data.table]{data.table}.}

\item{as_graph}{Return the object as a \link[tidygraph]{tbl_graph}.}

\item{as_granges}{Return the object as a \link[GenomicRanges]{GRanges}.}
}
\value{
Converted data.

Adjacency matrix
}
\description{
Functions to convert one object type to another.
}
\section{Functions}{
\itemize{
\item \code{dt_to_matrix()}: convert_

\item \code{graph_to_dt()}: convert_

\item \code{graph_to_ggnetwork()}: convert_

\item \code{ontology_to()}: convert_
Convert ontology

Convert an  \link[simona]{ontology_DAG} to
a number of other useful formats.

\item \code{ontology_to_adjacency()}: convert_
Adjacency matrix

Create adjacency matrix of HPO child-parent relationships.

\item \code{ontology_to_graph()}: convert_

\item \code{to_graph()}: convert_

}}
\examples{
dt <- data.table::as.data.table(mtcars, keep.rownames = TRUE)
X <- dt_to_matrix(dt)
ont <- get_ontology("hp", terms=10)
g <- ontology_to(ont, to="tbl_graph")
dat <- graph_to_dt(g)
g <- igraph::graph.atlas(10)
ggn <- graph_to_ggnetwork(g)
ont <- get_ontology()
obj <- ontology_to(ont=ont, to="dendrogram")
ont <- get_ontology()
adj <- ontology_to_adjacency(ont)
g <- igraph::graph.full(10)
g2 <- to_graph(g)
}
\concept{convert_}
\keyword{internal}
