% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_docs.R, R/get_metadata_omim.R,
%   R/get_metadata_orphanet.R, R/get_ontology_github.R, R/get_prevalence_oard.R,
%   R/get_prevalence_orphanet.R
\name{get_}
\alias{get_}
\alias{get_metadata_omim}
\alias{get_metadata_orphanet}
\alias{get_ontology_github}
\alias{get_prevalence_oard}
\alias{get_prevalence_orphanet}
\title{Get functions}
\usage{
get_metadata_omim(save_dir = cache_dir())

get_metadata_orphanet(save_dir = cache_dir())

get_ontology_github(
  file = "mondo-base.obo",
  repo = "monarch-initiative/mondo",
  tag = "latest",
  save_dir = cache_dir(),
  force_new = FALSE,
  ...
)

get_prevalence_oard(ids)

get_prevalence_orphanet(
  agg_by = c("MONDO_ID", "id", "Name"),
  include_mondo = TRUE
)
}
\arguments{
\item{save_dir}{Directory to save a file to.}

\item{file}{name or vector of names of files to be downloaded. If \code{NULL},
all assets attached to the release will be downloaded.}

\item{repo}{Repository name in format "owner/repo". Defaults to \code{guess_repo()}.}

\item{tag}{tag for the GitHub release to which this data should be attached.}

\item{force_new}{Create a new file instead of using any cached files.}

\item{...}{
  Arguments passed on to \code{\link[simona:import_obo]{simona::import_ontology}}
  \describe{
    \item{\code{verbose}}{Whether to print messages.}
    \item{\code{robot_jar}}{The path of the \code{robot.jar} file. It can be downloaded from https://github.com/ontodev/robot/releases.
Internally, the file is converted to the obo format and parsed by \code{import_obo()}. The value of \code{robot_jar} can be
set as a global option \code{simona_opt$robot_jar = ...}.}
    \item{\code{JAVA_ARGS}}{Options for \code{java}. For example you can set \code{-Xmx20G} if you want to increase the memory to 20G for java.}
  }}

\item{dat}{Input \link[data.table]{data.table}.}
}
\value{
Formatted data.

\link[simona]{ontology_DAG}
}
\description{
Functions to get data resources.
}
\section{Functions}{
\itemize{
\item \code{get_metadata_omim()}: get_

\item \code{get_metadata_orphanet()}: get_

\item \code{get_ontology_github()}: get_

Get ontology from GitHub

\item \code{get_prevalence_oard()}: get_

\item \code{get_prevalence_orphanet()}: get_

}}
\examples{
mondo <- get_mondo_ont()
}
\seealso{
Other get_: 
\code{\link{get_monarch_}},
\code{\link{get_upheno_}}
}
\concept{get_}
