% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alphamissense.R
\name{get_alphamissense}
\alias{get_alphamissense}
\title{Get AlphaMissense predictions}
\usage{
get_alphamissense(
  types = c("canonical", "non_canonical", "merged"),
  agg_fun = mean,
  save_dir = cache_dir(),
  force_new = FALSE
)
}
\arguments{
\item{types}{A character vector of types to return.}

\item{agg_fun}{A function to aggregate multiple transcripts per gene.}

\item{save_dir}{The directory to save the data to.}

\item{force_new}{If TRUE, force a new download.}

\item{verbose}{If TRUE, print messages.}
}
\value{
A named list of data.tables of AlphaMissense predictions.
}
\description{
Get gene-level
\href{https://console.cloud.google.com/storage/browser/dm_alphamissense}{
AlphaMissense} predictions for all canonical and non-canonical
protein-coding gene transcripts.
}
\examples{
\dontrun{
am <- get_alphamissense()
}
}
