% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ontology.R
\name{get_ontology}
\alias{get_ontology}
\title{Get ontology}
\usage{
get_ontology(
  name = c("mondo", "hpo", "upheno"),
  terms = NULL,
  filetype = "-base.obo",
  method = c("github", "bioportal"),
  add_metadata = TRUE,
  save_dir = cache_dir(),
  force_new = FALSE,
  ...
)
}
\arguments{
\item{name}{\itemize{
\item{"mondo"}{
Import the \href{https://mondo.monarchinitiative.org/}{Mondo} ontology.
\href{https://github.com/monarch-initiative/mondo/release}{
GitHub.}}.
\item{"hpo"}{
Import the \href{https://hpo.jax.org/app/}{Human Phenotype Ontology}.
\href{https://github.com/obophenotype/human-phenotype-ontology/release}{
GitHub.}}
}}

\item{add_metadata}{Add metadata to the resulting ontology object.}

\item{...}{
  Arguments passed on to \code{\link[=get_ontology_github]{get_ontology_github}}
  \describe{
    \item{\code{force_new}}{Create a new file instead of using any cached files.}
    \item{\code{save_dir}}{Directory to save a file to.}
    \item{\code{file}}{name or vector of names of files to be downloaded. If \code{NULL},
all assets attached to the release will be downloaded.}
    \item{\code{repo}}{Repository name in format "owner/repo". Defaults to \code{guess_repo()}.}
    \item{\code{tag}}{tag for the GitHub release to which this data should be attached.}
  }}
}
\value{
\link[simona]{ontology_DAG}
}
\description{
Import an up-to-date ontology directly from from the creators.
}
\examples{
mondo <- get_ontology(name="mondo")
hpo <- get_ontology(name="hpo")
upheno <- get_ontology(name="upheno")
}
