% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ontology_github.R
\name{get_ontology_github}
\alias{get_ontology_github}
\title{Get ontology from GitHub and import it via \link[simona]{import_ontology}.}
\usage{
get_ontology_github(
  name,
  repo,
  filetype = ".obo",
  file = paste0(name, filetype),
  tag = "latest",
  save_dir = cache_dir(),
  force_new = FALSE,
  ...
)
}
\arguments{
\item{name}{\itemize{
\item{<...>}{Any ontology name from \link{get_ols_options}}
\item{"mondo"}{
Import the \href{https://mondo.monarchinitiative.org/}{Mondo} ontology.
\href{https://github.com/monarch-initiative/mondo/release}{
GitHub.}}.
\item{"hpo"}{
Import the \href{https://hpo.jax.org/app/}{Human Phenotype Ontology}.
\href{https://github.com/obophenotype/human-phenotype-ontology/release}{
GitHub.}}
}}

\item{repo}{Repository name in format "owner/repo". Defaults to \code{guess_repo()}.}

\item{filetype}{File type to search for.}

\item{file}{Can be one of the following:
\itemize{
\item{"ontology"}{Creates an \link[simona]{ontology_DAG} R object by
 importing the OBO file directly from the official
 \href{https://github.com/obophenotype/upheno}{uPheno GitHub repository}.}
\item{"bestmatches"}{Returns a merged table with the best matches between
 human and non-human homologous phenotypes (from multiple species).
 Distributed by the official
 \href{https://github.com/obophenotype/upheno/tree/master/mappings}{
 uPheno GitHub repository}.}
\item{"upheno_mapping"}{Return a merged table with matches between human
and non-human homologous phenotypes (from multiple species).
Distributed by the
 \href{https://data.monarchinitiative.org/upheno2/current/upheno-release/all/index.html}{
 Monarch Initiative server}.}
}}

\item{tag}{tag for the GitHub release to which this data should be attached.}

\item{save_dir}{Directory to save a file to.}

\item{force_new}{If TRUE, force a new download.}

\item{...}{
  Arguments passed on to \code{\link[simona:import_obo]{simona::import_ontology}}
  \describe{
    \item{\code{verbose}}{Whether to print messages.}
    \item{\code{robot_jar}}{The path of the \code{robot.jar} file. It can be downloaded from https://github.com/ontodev/robot/releases.
Internally, the file is converted to the obo format and parsed by \code{import_obo()}. The value of \code{robot_jar} can be
set as a global option \code{simona_opt$robot_jar = ...}.}
    \item{\code{JAVA_ARGS}}{Options for \code{java}. For example you can set \code{-Xmx20G} if you want to increase the memory to 20G for java.}
  }}
}
\value{
\link[simona]{ontology_DAG}
}
\description{
Get ontology from GitHub and import it via \link[simona]{import_ontology}.
}
\keyword{internal}
