% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ontology_robot.R
\name{get_ontology_robot}
\alias{get_ontology_robot}
\title{Get ontology robot}
\usage{
get_ontology_robot(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_data]{get_data}}
  \describe{
    \item{\code{add_version}}{Add the release version
to the returned object's \link[base]{attributes}}
    \item{\code{file}}{Can be one of the following:
\itemize{
\item{"ontology"}{Creates an \link[simona]{ontology_DAG} R object by
 importing the OBO file directly from the official
 \href{https://github.com/obophenotype/upheno}{uPheno GitHub repository}.}
\item{"bestmatches"}{Returns a merged table with the best matches between
 human and non-human homologous phenotypes (from multiple species).
 Distributed by the official
 \href{https://github.com/obophenotype/upheno/tree/master/mappings}{
 uPheno GitHub repository}.}
\item{"upheno_mapping"}{Return a merged table with matches between human
and non-human homologous phenotypes (from multiple species).
Distributed by the
 \href{https://data.monarchinitiative.org/upheno2/current/upheno-release/all/index.html}{
 Monarch Initiative server}.}
}}
    \item{\code{save_dir}}{Directory to save a file to.}
    \item{\code{repo}}{Repository name in format "owner/repo". Defaults to \code{guess_repo()}.}
    \item{\code{tag}}{tag for the GitHub release to which this data should be attached.}
    \item{\code{overwrite}}{Should any local files of the same name be overwritten?
default \code{TRUE}.}
  }}
}
\description{
Download the ontology \href{https://github.com/ontodev/robot}{robot} tool
 and set its path so that it's accessible by \pkg{simona}.
}
\keyword{internal}
