% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_docs.R, R/get_upheno.R
\name{get_upheno_}
\alias{get_upheno_}
\alias{get_upheno}
\title{Get uPheno functions}
\usage{
get_upheno(file = c("ontology", "bestmatches", "upheno_mapping"))
}
\arguments{
\item{file}{Can be one of the following:
\itemize{
\item{"ontology"}{Creates an \link[simona]{ontology_DAG} R object by
 importing the OBO file directly from the official
 \href{https://github.com/obophenotype/upheno}{uPheno GitHub repository}.}
\item{"bestmatches"}{Returns a merged table with the best matches between
 human and non-human homologous phenotypes (from multiple species).
 Distributed by the official
 \href{https://github.com/obophenotype/upheno/tree/master/mappings}{
 uPheno GitHub repository}.}
\item{"upheno_mapping"}{Return a merged table with matches between human
and non-human homologous phenotypes (from multiple species).
Distributed by the
 \href{https://data.monarchinitiative.org/upheno2/current/upheno-release/all/index.html}{
 Monarch Initiative server}.}
}}
}
\value{
Formatted data.

\link[simona]{ontology_DAG} or \link[data.table]{data.table}.
}
\description{
Functions to get uPheno data resources.
}
\section{Functions}{
\itemize{
\item \code{get_upheno()}: get_ 
Get uPheno

Get data from the \href{https://github.com/obophenotype/upheno}{
Unified Phenotype Ontology (uPheno)}.

}}
\examples{
upheno <- get_upheno()
}
\seealso{
Other get_: 
\code{\link{get_monarch_}},
\code{\link{get_}}
}
\concept{get_}
