% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0docs.R
\name{main_}
\alias{main_}
\title{Main functions}
\arguments{
\item{ont}{An ontology of class \link[simona]{ontology_DAG}.}

\item{terms}{A vector of ontology term IDs.}

\item{g}{\link[tidygraph]{tbl_graph} object.}

\item{force_new}{Create a new file instead of using any cached files.}

\item{save_dir}{Directory to save a file to.}

\item{save_path}{File name to save to.}

\item{obj}{R object.}
}
\value{
R object.
}
\description{
Documentation for common arguments.
}
\concept{main_}
