% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_docs.R, R/map_monarch_genes.R, R/map_mondo.R
\name{map_}
\alias{map_}
\alias{map_monarch_genes}
\alias{map_mondo}
\title{Map functions}
\usage{
map_monarch_genes(dat, gene_col, all.x = FALSE)

map_mondo(
  dat,
  input_col = "disease_id",
  output_col = "MONDO_ID",
  to = "MONDO",
  map_types = NULL,
  map_to = "hpo",
  top_n = NULL,
  add_name = TRUE,
  add_definitions = TRUE,
  all.x = TRUE,
  allow.cartesian = FALSE
)
}
\arguments{
\item{dat}{\link[data.table]{data.table} with genes.}

\item{gene_col}{Name of the gene column in \code{dat}.}

\item{all.x}{logical; if \code{TRUE}, then extra rows will be added to the
output, one for each row in \code{x} that has no matching row in \code{y}.
These rows will have 'NA's in those columns that are usually filled with values
from \code{y}.  The default is \code{FALSE}, so that only rows with data from both
\code{x} and \code{y} are included in the output.}

\item{input_col}{Column name of input IDs.}

\item{output_col}{Column name of output IDs.}

\item{to}{Character vector of database(s) to map IDs to.
When not \code{"MONDO"}, can supply multiple alternative databases to map to
 (e.g. \code{c("OMIM","Orphanet","DECIPHER")}).}

\item{map_types}{Mapping types to include.}

\item{map_to}{Mapping outputs to include
(from Mondo IDs to another database's IDs).}

\item{top_n}{Top number of mappings to return per \code{top_by} grouping.
Set to \code{NULL} to skip this step.}

\item{add_name}{Logical, if TRUE, add MONDO name column.}

\item{add_definitions}{logical, if TRUE, add MONDO definition column.}

\item{allow.cartesian}{See \code{allow.cartesian} in \code{\link[data.table]{[.data.table}}.}

\item{top_by}{Grouping columns when selecting \code{top_n} rows per grouping.
Can be a character vector of one or more column names.}

\item{map_type_order}{The order in which \code{map_types} will be prioritised
when filtering the \code{top_n} rows by groupings.}

\item{map_orthologs}{Add gene-level data.}
}
\value{
Mapped data.

Mapped \code{dat}
}
\description{
Functions to map IDs across ontologies/databases.
}
\section{Functions}{
\itemize{
\item \code{map_monarch_genes()}: map_
Map Monarch genes

Map Monarch gene IDs to HGNC gene symbols, within or across species.

\item \code{map_mondo()}: map_
Map to/from MONDO IDs

}}
\examples{
dat <- example_dat("gene")
dt2 <- map_monarch_genes(dat=dat, gene_col="gene")
dat <- example_dat()
dat2 <- map_mondo(dat = dat)
}
\seealso{
Other map_: 
\code{\link{map_upheno_}}
}
\concept{map_}
