% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontology_to_adjacency.R
\name{ontology_to_adjacency}
\alias{ontology_to_adjacency}
\title{Adjacency matrix}
\usage{
ontology_to_adjacency(
  ont,
  terms = ont@terms,
  remove_terms = grep(":", terms, invert = TRUE, value = TRUE)
)
}
\arguments{
\item{terms}{Term IDs to include. Can alternatively be an integer, 
which will be used to randomly sample N terms from the data.}

\item{remove_terms}{Character vector of term IDs to exclude.}
}
\value{
Adjacency matrix
}
\description{
Create adjacency matrix of HPO child-parent relationships.
}
\examples{
ont <- get_ontology()
adj <- ontology_to_adjacency(ont)
}
