% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_upheno.R
\name{plot_upheno}
\alias{plot_upheno}
\title{Map uPheno plot}
\usage{
plot_upheno(
  pheno_map_genes_match = NULL,
  filters = list(db1 = "HP", gene_taxon_label1 = "Homo sapiens"),
  types = c("rainplot", "scatterplot", "heatmap")
)
}
\arguments{
\item{pheno_map_genes_match}{uPheno cross-species mapping data generated by
\link[KGExplorer]{map_upheno_data}.}

\item{types}{A character vector of plot types to generate.}

\item{subset_db1}{A character vector of reference ontologies to subset
\code{pheno_map_genes_match} by.}
}
\value{
A named list of \pkg{ggplot2} objects.
}
\description{
Generate multiple kinds of plots summarising mappings of phenotypes and genes
between the HPO and various non-human ontologies within uPheno.
}
\examples{
\dontrun{
pheno_map_genes_match <- map_upheno_data()
upheno_plots <- plot_upheno(
  pheno_map_genes_match = pheno_map_genes_match)
}
}
