% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_docs.R, R/query_monarch.R, R/query_oard.R
\name{query_}
\alias{query_}
\alias{query_monarch}
\alias{query_oard}
\title{Query functions}
\usage{
query_monarch(...)

query_oard(
  ids,
  concept_prefix = "q=",
  domain = "https://rare.cohd.io/api/",
  dataset_id = NULL,
  endpoint = "vocabulary/findConceptByAny",
  batch_size = 100,
  workers = 1
)
}
\value{
Queried data.
}
\description{
Functions to query specific subset of data via dedicated APIs.
}
\section{Functions}{
\itemize{
\item \code{query_monarch()}: query_

\item \code{query_oard()}: query_

}}
\concept{query_}
