% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_phenoscape.R
\name{query_phenoscape}
\alias{query_phenoscape}
\title{Query Phenoscape}
\source{
{https://github.com/phenoscape/rphenoscape/}{rphenoscape}
}
\usage{
query_phenoscape()
}
\description{
Query \href{https://phenoscape.org/}{Phenoscape}.
}
