% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_ubergraph.R
\name{query_ubergraph}
\alias{query_ubergraph}
\title{Query ubergraph}
\source{
\href{https://github.com/INCATools/ubergraph}{ubergraph}

\href{https://github.com/INCATools/ubergraph/tree/master/sparql}{
ubergraoh sparql query functions}

\href{https://github.com/phenoscape/ubeRsim}{ubeRsim}

\href{https://github.com/dhimmel/obonet}{obonet}
}
\usage{
query_ubergraph()
}
\description{
Query Ubergraph: an integrated multi-ontology knowledge graph.
}
