% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_dt.R
\name{unlist_dt}
\alias{unlist_dt}
\title{Unlist a \link{data.table}}
\usage{
unlist_dt(dat, drop = FALSE, collapse = ";")
}
\arguments{
\item{dat}{\link{data.table}}

\item{drop}{Drop columns that are lists.}

\item{collapse}{Character to collapse lists with.
Used only when drop is \code{FALSE}.}
}
\value{
\link{data.table}
}
\description{
\link{data.table}s can sometimes have columns that are nested lists. 
This function will either drop these columns or convert them into 
character strings.
}
\examples{
dat <- data.table::data.table(a=1:3,b=list(1:2,3:4))
unlist_dt(dat)
}
