% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_docs.R, R/add_db.R, R/cache_dir.R,
%   R/example_dat.R, R/filter_dt.R, R/get_data.R, R/messager.R, R/pkg_data.R,
%   R/report_missing.R, R/stopper.R
\name{utils_}
\alias{utils_}
\alias{add_db}
\alias{cache_dir}
\alias{example_dat}
\alias{filter_dt}
\alias{get_data}
\alias{messager}
\alias{pkg_data}
\alias{report_missing}
\alias{stopper}
\title{Utility functions}
\usage{
add_db(
  dat,
  input_col,
  output_col = paste0(input_col, "_db"),
  split = ":",
  keep = 1
)

cache_dir(package = "KGExplorer")

example_dat(
  types = c("omim_id", "orphanet_id", "decipher_id", "hpo_id", "gene")
)

filter_dt(dat, filters)

get_data(
  file,
  tag = "latest",
  repo = "neurogenomics/KGExplorer",
  save_dir = cache_dir(),
  add_version = FALSE,
  overwrite = TRUE
)

messager(..., v = Sys.getenv("VERBOSE") != "FALSE", parallel = FALSE)

pkg_data(name, package = "KGExplorer")

report_missing(dat, id_col, report_col)

stopper(..., v = TRUE)
}
\arguments{
\item{package}{a character string giving the name of an \R package}

\item{types}{Types of IDs to include in the data.}

\item{file}{name or vector of names of files to be downloaded. If \code{NULL},
all assets attached to the release will be downloaded.}

\item{tag}{tag for the GitHub release to which this data should be attached.}

\item{repo}{Repository name in format "owner/repo". Defaults to \code{guess_repo()}.}

\item{save_dir}{Directory to save data to.}

\item{add_version}{Add the release version
to the returned object's \link[base]{attributes}}

\item{overwrite}{Should any local files of the same name be overwritten?
default \code{TRUE}.}

\item{v}{Whether to print messages or not.}

\item{parallel}{Whether to enable message print when wrapped 
in parallelised functions.}
}
\value{
Merged data.

Cache directory path.

\link[data.table]{data.table}

Path to downloaded file or the object itself (when ".rds" format).

Null

Null
}
\description{
Functions to support other functions.
}
\section{Functions}{
\itemize{
\item \code{add_db()}: utils_

\item \code{cache_dir()}: utils_
Cache directory

Provides the path to the package-wide cache directory.

\item \code{example_dat()}: utils_
Example data

Construct an example \link[data.table]{data.table} containing diseases, 
phenotype, and/or genes.

\item \code{filter_dt()}: utils_

\item \code{get_data()}: utils_
Get remote data

Download remotely stored data via \link[piggyback]{pb_download}.

\item \code{messager()}: utils_
Print messages 

Conditionally print messages.
 Allows developers to easily control verbosity of functions, 
 and meet Bioconductor requirements that dictate the message 
 must first be stored to a variable before passing to \link[base]{message}.

\item \code{pkg_data()}: utils_

\item \code{report_missing()}: utils_

\item \code{stopper()}: utils_
Stop messages

Conditionally print stop messages.
Allows developers to easily control verbosity of functions, 
 and meet Bioconductor requirements that dictate the stop message 
 must first be stored to a variable before passing to \link[base]{stop}.

}}
\examples{
save_dir <- cache_dir
dat <- example_dat()
}
\concept{utils_}
\keyword{internal}
