% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_docs.R, R/add_ancestor.R
\name{add_}
\alias{add_}
\alias{add_ancestor}
\title{Add functions}
\usage{
add_ancestor(ont, lvl = 2, include_self = TRUE, force_new = FALSE)
}
\arguments{
\item{include_self}{For \code{dag_offspring()} and \code{dag_ancestors()}, this controls whether to also include the query term itself.}
}
\value{
Added data.
}
\description{
Functions to add extra metadata to an ontology or data.table object.
}
\section{Functions}{
\itemize{
\item \code{add_ancestor()}: add_
Add ancestor

For each term, get its ancestor at a given level (1 being the highest) 
and add the ID and name of the ancestor to the ontology metadata.

}}
\examples{
ont <- get_ontology("hpo", terms=10)
ont2 <- add_ancestor(ont)
}
\concept{add_}
