% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ontology_metadata.R
\name{add_ontology_metadata}
\alias{add_ontology_metadata}
\title{Add ontology metadata}
\usage{
add_ontology_metadata(
  ont,
  add_ancestors = 2,
  add_n_edges = TRUE,
  add_ontology_levels = TRUE
)
}
\arguments{
\item{ont}{An ontology of class \link[simona]{ontology_DAG}.}

\item{add_ancestors}{Add ancestors for each term.}

\item{add_n_edges}{Add the number of edges (connections) for each term.}

\item{add_ontology_levels}{Add the ontology level for each term.}
}
\description{
Add per-term metadata to ontology.
}
\examples{
ont <- get_ontology("hp", terms=10)
ont2 <- add_ontology_metadata(ont)
}
