% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_docs.R, R/cache_clear.R
\name{cache_}
\alias{cache_}
\alias{cache_clear}
\title{Cache functions}
\usage{
cache_clear(save_dir = cache_dir(), recursive = TRUE, ...)
}
\arguments{
\item{save_dir}{Path to cache directory.}

\item{recursive}{logical.  Should directories be deleted recursively?}

\item{...}{
  Arguments passed on to \code{\link[base:unlink]{base::unlink}}
  \describe{
    \item{\code{x}}{a character vector with the names of the file(s) or
    directories to be deleted.}
    \item{\code{force}}{logical.  Should permissions be changed (if possible) to
    allow the file or directory to be removed?}
    \item{\code{expand}}{logical. Should wildcards (see \sQuote{Details} below) and
    tilde (see \code{\link[base]{path.expand}}) be expanded?}
  }}
}
\value{
Null.

Null.
}
\description{
Functions to cache objects in order to speed up processes the second time.
}
\section{Functions}{
\itemize{
\item \code{cache_clear()}: utils_
Clear cache

Remove all data cached by the R package.

}}
\examples{
\dontrun{
cache_clear()
}
}
\concept{cache_}
