% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_docs.R, R/cache_clear.R, R/cache_files.R,
%   R/cache_save.R
\name{cache_}
\alias{cache_}
\alias{cache_clear}
\alias{cache_files}
\alias{cache_save}
\title{Cache functions}
\usage{
cache_clear(save_dir = cache_dir(), recursive = TRUE, ...)

cache_files(save_dir = cache_dir(), recursive = TRUE, ...)

cache_save(obj, path)
}
\arguments{
\item{save_dir}{Path to cache directory.}

\item{recursive}{logical.  Should directories be deleted recursively?}

\item{...}{
  Arguments passed on to \code{\link[base:unlink]{base::unlink}}, \code{\link[base:unlink]{base::unlink}}
  \describe{
    \item{\code{x}}{a character vector with the names of the file(s) or
    directories to be deleted.}
    \item{\code{force}}{logical.  Should permissions be changed (if possible) to
    allow the file or directory to be removed?}
    \item{\code{expand}}{logical. Should wildcards (see \sQuote{Details} below) and
    tilde (see \code{\link[base]{path.expand}}) be expanded?}
  }}

\item{obj}{R object.}

\item{path}{File name to save to.}
}
\value{
Null.

Null.
}
\description{
Functions to cache objects in order to speed up processes the second time.
}
\section{Functions}{
\itemize{
\item \code{cache_clear()}: cache_
Clear cache

Remove all data cached by the R package.

\item \code{cache_files()}: cache_
List cache

List all data cached by the R package.

\item \code{cache_save()}: cache_
Save cache.

}}
\examples{
\dontrun{
cache_clear()
}
\dontrun{
cache_files()
}
cache_save(mtcars, tempfile())
}
\concept{cache_}
