% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_docs.R, R/graph_to_dt.R, R/ontology_to.R,
%   R/ontology_to_tidygraph.R, R/to_tidygraph.R
\name{convert_}
\alias{convert_}
\alias{graph_to_dt}
\alias{ontology_to}
\alias{ontology_to_tidygraph}
\alias{to_tidygraph}
\title{Converter functions}
\usage{
graph_to_dt(g, what = c("nodes", "edges"), id_col = c("id", "name")[1])

ontology_to(
  ont,
  terms = ont@terms,
  remove_terms = grep(":", terms, invert = TRUE, value = TRUE),
  to = c("adjacency", "adjacency_dist", "adjacency_dist_hclust",
    "adjacency_dist_hclust_dendrogram", "adjacency_dist_hclust_clusters", "similarity",
    "dendrogram", "igraph", "dot", "adjacency_igraph", "igraph_dist",
    "igraph_dist_hclust", "igraph_dist_hclust_dendrogram", "tidygraph", "data.frame",
    "data.table", "list"),
  as_sparse = FALSE,
  ...
)

ontology_to_tidygraph(ont, ...)

to_tidygraph(g, ...)
}
\arguments{
\item{ont}{An ontology of class \link[simona]{ontology_DAG}.}

\item{terms}{Term IDs to include. Can alternatively be an integer, 
which will be used to randomly sample N terms from the data.}

\item{remove_terms}{Character vector of term IDs to exclude.}

\item{to}{A character string specifying the format to convert to.}

\item{as_sparse}{If TRUE, return a sparse matrix where possible.}

\item{...}{Additional arguments passed to plot-specific functions.}
}
\value{
Converted data.
}
\description{
Functions to convert one object type to another.
}
\section{Functions}{
\itemize{
\item \code{graph_to_dt()}: convert_

\item \code{ontology_to()}: convert_
Convert ontology

Convert an  \link[simona]{ontology_DAG} to
a number of other useful formats.

\item \code{ontology_to_tidygraph()}: convert_

\item \code{to_tidygraph()}: convert_

}}
\examples{
g <- igraph::graph.atlas(10)
dat <- graph_to_dt(g)
ont <- get_ontology()
obj <- ontology_to(ont=ont, to="dendrogram")
g <- igraph::graph.full(10)
g2 <- to_tidygraph(g)
}
\concept{convert_}
\keyword{internal}
