% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_docs.R, R/filter_ontology.R
\name{filter_}
\alias{filter_}
\alias{filter_ontology}
\title{Filter functions}
\usage{
filter_ontology(
  ont,
  terms = NULL,
  remove_terms = NULL,
  use_simona = FALSE,
  ...
)
}
\arguments{
\item{terms}{Term IDs to include. Can alternatively be an integer, 
which will be used to randomly sample N terms from the data.}

\item{remove_terms}{Character vector of term IDs to exclude.}

\item{use_simona}{Use \link[simona]{dag_filter} to filter terms.}
}
\value{
Converted data.
}
\description{
Functions to filter objects
}
\section{Functions}{
\itemize{
\item \code{filter_ontology()}: filter_
Filter ontology

Filter ontology by terms.

}}
\examples{
ont <- get_ontology("hp")
ont2 <- filter_ontology(ont,terms=c("HP:0000001","HP:0000002"))
ont3 <- filter_ontology(ont,terms=100)
}
\concept{filter_}
