% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_docs.R, R/get_metadata_omim.R,
%   R/get_metadata_orphanet.R, R/get_ontology_github.R, R/get_prevalence_oard.R,
%   R/get_prevalence_orphanet.R, R/get_upheno.R, R/get_version.R
\name{get_}
\alias{get_}
\alias{get_metadata_omim}
\alias{get_metadata_orphanet}
\alias{get_ontology_github}
\alias{get_prevalence_oard}
\alias{get_prevalence_orphanet}
\alias{get_upheno}
\alias{get_version}
\title{Get functions}
\usage{
get_metadata_omim(save_dir = cache_dir())

get_metadata_orphanet(save_dir = cache_dir())

get_ontology_github(
  file = "mondo-base.obo",
  repo = "monarch-initiative/mondo",
  tag = "latest",
  save_dir = cache_dir(),
  force_new = FALSE,
  ...
)

get_prevalence_oard(ids)

get_prevalence_orphanet(
  agg_by = c("MONDO_ID", "id", "Name"),
  include_mondo = TRUE
)

get_upheno(file = c("ontology", "bestmatches", "upheno_mapping"))

get_version(obj, return_version = FALSE, verbose = TRUE)
}
\arguments{
\item{save_dir}{Directory to save a file to.}

\item{file}{Can be one of the following:
\itemize{
\item{"ontology"}{Creates an \link[simona]{ontology_DAG} R object by
 importing the OBO file directly from the official
 \href{https://github.com/obophenotype/upheno}{uPheno GitHub repository}.}
\item{"bestmatches"}{Returns a merged table with the best matches between
 human and non-human homologous phenotypes (from multiple species).
 Distributed by the official
 \href{https://github.com/obophenotype/upheno/tree/master/mappings}{
 uPheno GitHub repository}.}
\item{"upheno_mapping"}{Return a merged table with matches between human
and non-human homologous phenotypes (from multiple species).
Distributed by the
 \href{https://data.monarchinitiative.org/upheno2/current/upheno-release/all/index.html}{
 Monarch Initiative server}.}
}}

\item{repo}{Repository name in format "owner/repo". Defaults to \code{guess_repo()}.}

\item{tag}{tag for the GitHub release to which this data should be attached.}

\item{force_new}{Create a new file instead of using any cached files.}

\item{...}{
  Arguments passed on to \code{\link[simona:import_obo]{simona::import_ontology}}
  \describe{
    \item{\code{robot_jar}}{The path of the \code{robot.jar} file. It can be downloaded from https://github.com/ontodev/robot/releases.
Internally, the file is converted to the obo format and parsed by \code{import_obo()}. The value of \code{robot_jar} can be
set as a global option \code{simona_opt$robot_jar = ...}.}
    \item{\code{JAVA_ARGS}}{Options for \code{java}. For example you can set \code{-Xmx20G} if you want to increase the memory to 20G for java.}
  }}

\item{obj}{An object.}

\item{return_version}{Return the version as a character string.}

\item{verbose}{Print messages.}

\item{dat}{Input \link[data.table]{data.table}.}

\item{filters}{A named list, where each element in the list is the name of 
a column in the data, and the vector within each element represents the 
values to include in the final data.}

\item{maps}{A list of paired to/from types to filter Monarch association 
files by. For example, \code{list(c("gene","disease"))} will return any 
 files that contains gene-disease associations.
Passes to \link{get_monarch_files}.}

\item{queries}{A list of free-form substring queries to filter files by 
(using any column in the metadata). 
For example, \code{list("gene_disease","variant_disease")} will return any 
 files that contain either of the substrings 
 "gene_disease" or "variant_disease".
Passes to \link{get_monarch_files}.}

\item{domain}{Web domain to search for Monarch files.}

\item{subdir}{Subdirectory path to search for Monarch files within
\code{domain}.}

\item{rbind}{If \code{TRUE}, rbinds all \link[data.table]{data.table}s 
together. Otherwise, returns a named list of separated 
\link[data.table]{data.table}s.}
}
\value{
Data..

\link[simona]{ontology_DAG}

\link[simona]{ontology_DAG} or \link[data.table]{data.table}.

HPO Release version a character string.
}
\description{
Functions to get data resources.
}
\section{Functions}{
\itemize{
\item \code{get_metadata_omim()}: get_

\item \code{get_metadata_orphanet()}: get_

\item \code{get_ontology_github()}: get_

Get ontology from GitHub

\item \code{get_prevalence_oard()}: get_

\item \code{get_prevalence_orphanet()}: get_

\item \code{get_upheno()}: get_ 
Get uPheno

Get data from the \href{https://github.com/obophenotype/upheno}{
Unified Phenotype Ontology (uPheno)}.

\item \code{get_version()}: get_
Get version

Extract the precise version of the HPO release that the data object
was built from. All HPO Releases can be found at the official
\href{https://github.com/obophenotype/human-phenotype-ontology/releases}{
HPO GitHub Releases page}.

}}
\examples{
mondo <- get_mondo_ont()
upheno <- get_upheno()
obj <- get_hpo()
get_version(obj=obj)
}
\concept{get_}
