% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ontology_robot.R
\name{get_ontology_robot}
\alias{get_ontology_robot}
\title{Get ontology robot}
\usage{
get_ontology_robot(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=get_data]{get_data}}
  \describe{
    \item{\code{save_dir}}{Directory to save data to.}
    \item{\code{add_version}}{Add the release version
to the returned object's \link[base]{attributes}}
    \item{\code{file}}{name or vector of names of files to be downloaded. If \code{NULL},
all assets attached to the release will be downloaded.}
    \item{\code{repo}}{Repository name in format "owner/repo". Defaults to \code{guess_repo()}.}
    \item{\code{tag}}{tag for the GitHub release to which this data should be attached.}
    \item{\code{overwrite}}{Should any local files of the same name be overwritten?
default \code{TRUE}.}
  }}
}
\description{
Download the ontology \href{https://github.com/ontodev/robot}{robot} tool
 and set its path so that it's accessible by \pkg{simona}.
}
\keyword{internal}
