% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_opentargets.R
\name{get_opentargets}
\alias{get_opentargets}
\title{Get OpenTargets}
\source{
\href{https://community.opentargets.org/t/r-script-for-graphql-query-query-targetdiseaseevidence/662/5}{OpenTargets GraphQL queries in R}
}
\usage{
get_opentargets(
  release = "latest",
  data_type = c("associationByDatasourceDirect", "associationByDatasourceIndirect",
    "associationByDatatypeDirect", "associationByDatatypeIndirect",
    "associationByOverallDirect", "associationByOverallIndirect")[1],
  server = "https://ftp.ebi.ac.uk/pub/databases/opentargets/platform/",
  ftp = paste0(server, release, "/output/etl/parquet/", data_type[1], "/"),
  save_dir = cache_dir(),
  force_new = FALSE
)
}
\description{
Get OpenTargets disease-gene associations data
}
\examples{
d <- get_opentargets()
}
