% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pli.R
\name{get_pli}
\alias{get_pli}
\title{Get pLI}
\usage{
get_pli(
  save_dir = tools::R_user_dir(package = "MultiEWCE", which = "cache"),
  agg_fun = mean,
  force_new = FALSE,
  verbose = TRUE
)
}
\value{
A data.table with the following columns
}
\description{
Get gene-level \href{https://gnomad.broadinstitute.org/faq#what-is-pli}{
pLI} scores for all canonical and non-canonical protein-coding gene
transcripts.
NOTE: The MANE Select set consists of one transcript at each protein-coding
locus across the genome that is representative of biology at that locus.
NOTE: Mapping genes with \link[orthogene]{map_genes} only reduces the number
of mapped genes compared to the provided "gene" column.
}
