% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0docs.R, R/plot_clinvar.R,
%   R/plot_ggnetwork.R, R/plot_graph_3d.R, R/plot_graph_visnetwork.R,
%   R/plot_ontology.R, R/plot_ontology_circular.R, R/plot_ontology_graphviz.R,
%   R/plot_ontology_heatmap.R, R/plot_ontology_tidygraph.R,
%   R/plot_ontology_visnetwork.R, R/plot_save.R, R/plot_ttd.R,
%   R/plot_upheno_heatmap.R
\name{plot_}
\alias{plot_}
\alias{plot_clinvar}
\alias{plot_ggnetwork}
\alias{plot_graph_3d}
\alias{plot_graph_visnetwork}
\alias{plot_ontology}
\alias{plot_ontology_circular}
\alias{plot_ontology_graphviz}
\alias{plot_ontology_heatmap}
\alias{plot_ontology_tidygraph}
\alias{plot_ontology_visnetwork}
\alias{plot_save}
\alias{plot_ttd}
\alias{plot_upheno_heatmap}
\title{Plot functions}
\source{
\href{https://www.r-bloggers.com/2013/09/network-visualization-part-4-3d-networks/}{
R bloggers}

\href{https://community.plotly.com/t/is-it-possible-to-connect-scatters-in-3d-scatter-plot/}{
Plotly: Connect points in 3D plot}

\href{https://plotly.com/r/line-charts/}{
Plotly: Connect points in 2D plot}

\href{https://plotly.com/r/3d-line-plots/}{
Plotly: 3D lines plots}

\href{https://plotly.com/python/v3/3d-network-graph/}{
Plotly: 3D network plot}
}
\usage{
plot_clinvar(
  hits,
  x = "region",
  y = "N",
  fill = "Type",
  by = c("build", "region", "Type", "ClinSigSimple"),
  rows = "ClinSigSimple"
)

plot_ggnetwork(
  g,
  colour_var = "x",
  size_var = colour_var,
  label_var = "name",
  hoverbox_column = "hover",
  interactive = TRUE,
  show_plot = TRUE,
  ...
)

plot_graph_3d(
  g,
  layout_func = igraph::layout.fruchterman.reingold,
  id_var = "name",
  node_color_var = "ancestor_name",
  edge_color_var = "zend",
  text_color_var = node_color_var,
  node_symbol_var = "ancestor_name",
  node_palette = pals::kovesi.cyclic_mrybm_35_75_c68_s25,
  edge_palette = pals::kovesi.cyclic_mrybm_35_75_c68_s25,
  node_opacity = 0.75,
  edge_opacity = 0.5,
  kde_palette = pals::gnuplot,
  add_kde = TRUE,
  extend_kde = 1.5,
  bg_color = kde_palette(6)[1],
  add_labels = FALSE,
  keep_grid = FALSE,
  aspectmode = "cube",
  hover_width = 100,
  label_width = 100,
  seed = 2023,
  showlegend = TRUE,
  show_plot = TRUE,
  save_path = tempfile(fileext = "plot_graph_3d.html"),
  verbose = TRUE
)

plot_graph_visnetwork(
  g,
  label_var = "name",
  size_var = "degree",
  colour_var = size_var,
  invert_colour_var = TRUE,
  columns = get_graph_colnames(g),
  preferred_palettes = NULL,
  selectedBy = label_var,
  show_plot = TRUE,
  layout = "layout_with_kk",
  solver = "forceAtlas2Based",
  physics = FALSE,
  forceAtlas2Based = list(avoidOverlap = 0.5, gravitationalConstant = -50),
  scaling = NULL,
  smooth = list(enabled = TRUE, type = "cubicBezier", roundness = 0.5),
  add_visExport = FALSE,
  degree = 1,
  width = "100\%",
  height = "90vh",
  highlight_color = "#00FFFFCF",
  randomSeed = 2024,
  main = NULL,
  submain = NULL,
  save_path = tempfile(fileext = "_visnetwork.html")
)

plot_ontology(
  ont,
  terms = NULL,
  types = c("circular", "graphviz", "tidygraph", "visnetwork"),
  ...
)

plot_ontology_circular(ont, ...)

plot_ontology_graphviz(ont, ...)

plot_ontology_heatmap(
  ont,
  annot = data.table::data.table(ont@elementMetadata),
  X = ontology_to(ont, to = "similarity"),
  fontsize = ont@n_terms * 4e-04,
  row_labels = ont@terms,
  column_labels = row_labels,
  name = NULL,
  row_side_vars = c("ancestor_name"),
  col_side_vars = c("IC", "depth", "n_children", "n_offspring", "n_connected_leaves"),
  col = pals::gnuplot(),
  show_plot = TRUE,
  save_path = tempfile(fileext = "plot_ontology_heatmap.pdf"),
  height = 12,
  width = height * 1.1,
  seed = 2023,
  types = c("heatmaply", "ComplexHeatmap")[2],
  ...
)

plot_ontology_tidygraph(ont, ...)

plot_ontology_visnetwork(ont, ...)

plot_save(plt, path, width = 7, height = width)

plot_ttd(dat_sub)

plot_upheno_heatmap(
  plot_dat,
  ont,
  hpo_ids = NULL,
  value.var = c("phenotype_genotype_score", "prop_intersect", "equivalence_score",
    "subclass_score"),
  min_rowsums = NULL,
  cluster_from_ontology = FALSE,
  save_dir = tempdir(),
  height = 15,
  width = 10
)
}
\arguments{
\item{x}{: Number. Default to undefined. This gives a node an initial x position. When using the hierarchical layout, either the x or y position is set by the layout engine depending on the type of view. The other value remains untouched. When using stabilization, the stabilized position may be different from the initial one. To lock the node to that position use the physics or fixed options.}

\item{y}{: Number. Default to undefined. This gives a node an initial y position. When using the hierarchical layout, either the x or y position is set by the layout engine depending on the type of view. The other value remains untouched. When using stabilization, the stabilized position may be different from the initial one. To lock the node to that position use the physics or fixed options.}

\item{g}{ggnetwork object 
(or an igraph/tbl_graph to be converted to ggnetwork format).}

\item{colour_var}{Column to color nodes by.}

\item{size_var}{Column to scale node size by.}

\item{label_var}{Column containing the label for each node in a graph
(e.g. "hpo_name").}

\item{interactive}{Make the plot interactive.}

\item{show_plot}{Print the plot after it's been generated.}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:aes]{ggplot2::aes}}, \code{\link[simona:dag_viz]{simona::dag_circular_viz}}, \code{\link[simona:dag_viz]{simona::dag_graphviz}}, \code{\link[base:plot]{base::plot}}, \code{\link[simona:dag_viz]{simona::dag_graphviz}}
  \describe{
    \item{\code{dag}}{An \code{ontology_Dag} object.}
    \item{\code{highlight}}{A vector of terms to be highlighted on the DAG.}
    \item{\code{start}}{Start of the circle, measured in degree.}
    \item{\code{end}}{End of the circle, measured in degree.}
    \item{\code{partition_by_level}}{If \code{node_col} is not set, users can cut the DAG into clusters with different node colors. The partitioning is applied by \code{\link[simona:partition_by_level]{partition_by_level()}}.}
    \item{\code{partition_by_size}}{Similar as \code{partition_by_level}, but the partitioning is applied by \code{\link[simona:partition_by_size]{partition_by_size()}}.}
    \item{\code{node_col}}{Colors of nodes. If the value is a vector, the order should correspond to terms in \code{\link[simona:dag_all_terms]{dag_all_terms()}}.}
    \item{\code{node_transparency}}{Transparency of nodes. The same format as \code{node_col}.}
    \item{\code{node_size}}{Size of nodes. The same format as \code{node_col}.}
    \item{\code{edge_col}}{A named vector where names correspond to relation types.}
    \item{\code{edge_transparency}}{A named vector where names correspond to relation types.}
    \item{\code{legend_labels_from}}{If partitioning is applied on the DAG, a legend is generated showing different top
terms. By default, the legend labels are the term IDs. If there are additionally column stored
in the meta data frame of the DAG object, the column name can be set here to replace the term IDs as
legend labels.}
    \item{\code{legend_labels_max_width}}{Maximal width of legend labels measured by the number of characters per line. Labels are wrapped into
multiple lines if the widths exceed it.}
    \item{\code{other_legends}}{A list of legends generated by \code{\link[ComplexHeatmap:Legend]{ComplexHeatmap::Legend()}}.}
    \item{\code{use_raster}}{Whether to first write the circular image into a temporary png file, then add to the plot
as a raster object?}
    \item{\code{newpage}}{Whether call \code{\link[grid:grid.newpage]{grid::grid.newpage()}} to create a new plot?}
    \item{\code{node_param}}{A list of parameters. Each parameter has the same format. The value can be
a single scalar, a full length vector with the same order as in \code{\link[simona:dag_all_terms]{dag_all_terms()}},
or a named vector that contains a subset of terms that need to be customized.
The full set of parameters can be found at \url{https://graphviz.org/docs/nodes/}.}
    \item{\code{edge_param}}{A list of parameters. Each parameter has the same format. The value can be a single
scalar, or a named vector that contains a subset of terms that need to be customized.
The full set of parameters can be found at \url{https://graphviz.org/docs/edges/}.
If the parameter is set to a named vector, it can be named by relation types \code{c("is_a" = ...)},
or directly relations \code{c("a -> b" = ...)}. Please see the vignette for details.}
  }}

\item{layout_func}{Layout function for the graph.}

\item{node_color_var}{Variable in the vertex metadata to color nodes by.}

\item{edge_color_var}{Variable in the edge metadata to color edges by.}

\item{text_color_var}{Variable in the node metadata to color text by.}

\item{node_symbol_var}{Variable in the vertex metadata to shape nodes by.}

\item{node_palette}{Color palette function for the nodes/points.}

\item{edge_palette}{Color palette function for the edges/lines.}

\item{node_opacity}{Node opacity.}

\item{edge_opacity}{Edge opacity.}

\item{kde_palette}{Color palette function for the KDE plot.}

\item{add_kde}{Add a kernel density estimation (KDE) plot
below the 3D scatter plot (i.e. the "mountains" beneath the points).}

\item{extend_kde}{Extend the area that the KDE plot covers.}

\item{bg_color}{Plot background color.}

\item{add_labels}{Add phenotype name labels to each point.}

\item{keep_grid}{Keep all grid lines and axis labels.}

\item{aspectmode}{The proportions of the 3D plot. See the
\href{https://plotly.com/python/reference/layout/scene/#layout-scene-aspectmode}{
plotly documentation site} for details.}

\item{hover_width}{Maximum width of the hover text.}

\item{label_width}{Maximum width of the label text.}

\item{seed}{Set the seed for reproducible clustering.}

\item{showlegend}{Show node fill legend.}

\item{save_path}{Path to save interactive plot to
as a self-contained HTML file.}

\item{verbose}{Print messages.}

\item{columns}{Names of columns to map colour palettes to.}

\item{preferred_palettes}{Preferred palettes to use for each column.}

\item{selectedBy}{: Custom option. Character or a named list. Add a multiple selection based on column of node data.frame creating an HTML select element. Not available for DOT and Gephi.
\itemize{
 \item{"variable"}{ : Character. Column name of selection variable.}
 \item{"values}{ : Optional. Vector of possible values. Default to all values in nodes data.frame.}
 \item{"selected"}{ : Optional. Integer/Character. Initial selection. Default to NULL}
 \item{"style"}{ : Optional. Character. HTML style of list. Default to 'width: 150px; height: 26px'.}
 \item{"multiple"}{ : Optional. Boolean. Default to FALSE. If TRUE, you can affect multiple groups per nodes using a comma ("gr1,gr2")}
 \item{"hideColor"}{ : Optional. String. Color for hidden nodes/edges. Use a rgba definition. Default to rgba(200,200,200,0.5)}
 \item{"main"}{ : Optional. Default to "Select by variable"}
 \item{"sort"}{ : Optional. If values is NULL, sort all possible values ?. Default to TRUE}
 \item{"highlight"}{ : Optional. Boolean. Run highlightNearest if defined on each selected node ? Default to FALSE}
}}

\item{layout}{: Character Name of igraph layout function to use. Default to "layout_nicely"}

\item{solver}{: String. Default to 'barnesHut'.    You can select your own solver. Possible options: 'barnesHut', 'repulsion', 'hierarchicalRepulsion', 'forceAtlas2Based'. When setting the hierarchical layout, the hierarchical repulsion solver is automaticaly selected, regardless of what you fill in here.}

\item{physics}{: Boolean. Default to FALSE. Enabled physics on nodes ?}

\item{forceAtlas2Based}{named list of options
\itemize{
 \item{"theta"}{ : Number. Default to 0.5. This parameter determines the boundary between consolidated long range forces and individual short range forces. To oversimplify higher values are faster but generate more errors, lower values are slower but with less errors.}
 \item{"gravitationalConstant"}{ : Number. Default to -50. Gravity attracts. We like repulsion. So the value is negative. If you want the repulsion to be stronger, decrease the value (so -10000, -50000).}
 \item{"centralGravity"}{ : Number. Default to 0.01. There is a central gravity attractor to pull the entire network back to the center.}
 \item{"springLength"}{ : Number. Default to 100. The edges are modelled as springs. This springLength here is the the rest length of the spring.}
 \item{"springConstant"}{ : Number. Default to 0.08. This is how 'sturdy' the springs are. Higher values mean stronger springs.}
 \item{"damping"}{ : Number. Default to 0.4. Accepted range: [0 .. 1]. The damping factor is how much of the velocity from the previous physics simulation iteration carries over to the next iteration.}
 \item{"avoidOverlap"}{ : Number. Default to 0. Accepted range: [0 .. 1]. When larger than 0, the size of the node is taken into account. The distance will be calculated from the radius of the encompassing circle of the node for both the gravity model. Value 1 is maximum overlap avoidance.}
}}

\item{scaling}{: Named list.  If the value option is specified, the size of the nodes will be scaled according to the properties in this object.
\itemize{
 \item{"min"}{ :  Number. Default to 10. If nodes have a value, their sizes are determined by the value, the scaling function and the min max values.}
 \item{"max"}{ : Number. Default to 30. This is the maximum allowed size when the nodes are scaled using the value option.}
 \item{"label"}{ : Named list or Boolean. Default to Named list. This can be false if the label is not allowed to scale with the node. If true it will scale using default settigns. For further customization, you can supply an object.
   \itemize{
     \item{"enabled"}{ : Boolean. Default to false. Toggle the scaling of the label on or off. If this option is not defined, it is set to true if any of the properties in this object are defined.}
     \item{"min"}{ : Number. Default to 14. The minimum font-size used for labels when scaling.}
     \item{"max"}{ : Number. Default to 30. The maximum font-size used for labels when scaling.}
     \item{"maxVisible"}{ :   Number. Default to 30. When zooming in, the font is drawn larger as well. You can limit the perceived font size using this option. If set to 30, the font will never look larger than size 30 zoomed at 100\%.}
     \item{"drawThreshold"}{ : Number. Default to 5. When zooming out, the font will be drawn smaller. This defines a lower limit for when the font is drawn. When using font scaling, you can use this together with the maxVisible to first show labels of important nodes when zoomed out and only show the rest when zooming in.}
   }
 }
 \item{"customScalingFunction"}{ : Function. If nodes have value fields, this function determines how the size of the nodes are scaled based on their values.}
}}

\item{smooth}{: Boolean. Default to FALSE. Use smooth edges ?}

\item{add_visExport}{Add PDF download button.}

\item{width}{: Number. Default to 1. The width of the edge. If value is set, this is not used.}

\item{height}{: String. Default to "100\%". The height of the network in pixels or as a percentage.}

\item{randomSeed}{: Number. The nodes are randomly positioned initially. This means that the settled result is different every time. If you provide a random seed manually, the layout will be the same every time.}

\item{main}{: For add a title. Character or a named list.
\itemize{
 \item{"text"}{ : Character. Title.}
 \item{"style"}{ : Optional. Character. HTML style of title. Default to 'font-family:Georgia, Times New Roman, Times, serif;font-weight:bold;font-size:20px;text-align:center;'.}
}}

\item{submain}{: For add a subtitle. Character or a named list.
\itemize{
 \item{"text"}{ : Character. Subtitle.}
 \item{"style"}{ : Optional. Character. HTML style of submain. Default to 'font-family:Georgia, Times New Roman, Times, serif;font-size:12px;text-align:center;'.}
}}

\item{ont}{An ontology of class \link[simona]{ontology_DAG}.}

\item{terms}{A subset of HPO IDs to include in the final dataset and plots
(e.g. c("HP:0001508","HP:0001507")).}

\item{types}{Types of graph to produce. Can be one or more.}

\item{fontsize}{Axis labels font size.}

\item{row_labels}{Optional row labels which are put as row names in the heatmap.}

\item{column_labels}{Optional column labels which are put as column names in the heatmap.}

\item{name}{Name of the heatmap. By default the heatmap name is used as the title of the heatmap legend.}

\item{row_side_vars}{Variables to include in row-side
metadata annotations.}

\item{col_side_vars}{Variables to include in column-side
metadata annotations.}

\item{col}{A vector of colors if the color mapping is discrete or a color mapping  function if the matrix is continuous numbers (should be generated by \code{\link[circlize]{colorRamp2}}). If the matrix is continuous, the value can also be a vector of colors so that colors can be interpolated. Pass to \code{\link[ComplexHeatmap]{ColorMapping}}. For more details and examples, please refer to \url{https://jokergoo.github.io/ComplexHeatmap-reference/book/a-single-heatmap.html#colors} .}

\item{save_dir}{Directory to save cached data.}

\item{id_col}{Column containing the unique identifier for each node 
in a graph (e.g. "name").}

\item{top_targets}{\link[data.table]{data.table} of prioritised targets
generated by \link[MultiEWCE]{prioritise_targets}.}
}
\value{
A named list containing the plot and the data.

A 3D interactive \link[plotly]{plotly} object.

Null

Null

Plot

Null

Null
}
\description{
Functions to plot data.
}
\section{Functions}{
\itemize{
\item \code{plot_clinvar()}: plot_
Plot mapped variant annotations.

\item \code{plot_ggnetwork()}: plot_
Plot a ggnetwork.

\item \code{plot_graph_3d()}: plot_
3D network

Plot a subset of the HPO as a 3D network.

\item \code{plot_graph_visnetwork()}: plot_
Plot graph using visNetwork.

\item \code{plot_ontology()}: plot_

\item \code{plot_ontology_circular()}: plot_

\item \code{plot_ontology_graphviz()}: plot_
Plot ontology: graphviz

Make a circular plot of an ontology.

\item \code{plot_ontology_heatmap()}: plot_
Plot heatmap

Plot a phenotype x phenotype correlation matrix based on genetic overlap.

\item \code{plot_ontology_tidygraph()}: plot_

\item \code{plot_ontology_visnetwork()}: plot_

\item \code{plot_save()}: plot_
Save a plot using \pkg{grDevices} or \link[visNetwork]{visSave}.

\item \code{plot_ttd()}: plot_
Plot TTD.

\item \code{plot_upheno_heatmap()}: plot_

}}
\examples{
ont <- get_ontology(terms=10)
g <- ontology_to(ont, to="tbl_graph")
g <- add_hoverboxes(g)
out <- plot_ggnetwork(g, label_var="label")
ont <- get_ontology("hp", terms=10, add_ancestors=TRUE)
g <- ontology_to(ont, to="tbl_graph")
plt <- plot_graph_3d(g=g, show_plot=FALSE)
ont <- get_ontology("hp", terms=10)
g <- ontology_to(ont, to="tbl_graph")
out <- plot_graph_visnetwork(g)
ont <- get_ontology("hp")
res <- plot_ontology(ont, 
                     terms=100,
                     types="circular",
                     partition_by_level=2,
                     edge_transparency=.9)
ont <- get_ontology("hp", terms=2)
hm <- plot_ontology_heatmap(ont)
}
\concept{plot_}
\keyword{internal}
