% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_upheno.R
\name{plot_upheno}
\alias{plot_upheno}
\title{Map uPheno plot}
\usage{
plot_upheno(
  pheno_map_genes_match = NULL,
  filters = list(db1 = "HP", gene_taxon_label1 = "Homo sapiens"),
  types = c("rainplot", "scatterplot", "heatmap")
)
}
\arguments{
\item{pheno_map_genes_match}{uPheno cross-species mapping data generated by
\link[KGExplorer]{map_upheno_data}.
\code{pheno_map_genes_match} by.}

\item{filters}{A named list, where each element in the list is the name of 
a column in the data, and the vector within each element represents the 
values to include in the final data.}

\item{types}{A character vector of plot types to generate.}
}
\value{
A named list of \pkg{ggplot2} objects.
}
\description{
Generate multiple kinds of plots summarising mappings of phenotypes and genes
between the HPO and various non-human ontologies within uPheno.
}
\examples{
\dontrun{
pheno_map_genes_match <- map_upheno_data()
upheno_plots <- plot_upheno(
  pheno_map_genes_match = pheno_map_genes_match)
}
}
