% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_ancestors.R
\name{prune_ancestors}
\alias{prune_ancestors}
\title{Prune ancestor}
\usage{
prune_ancestors(dat, id_col, ont)
}
\arguments{
\item{dat}{A \link{data.table} with a column of ontology terms.}

\item{id_col}{The name of the column containing ontology term IDs.}

\item{ont}{An ontology of class \link[simona]{ontology_DAG}.}
}
\description{
Prune redundant ancestral terms from a \link{data.table}.
}
\examples{
dat <- data.table::data.table(hpo_id=c("HP:0000001","HP:0000002","HP:0000003"),
                             name=c("term1","term2","term3"))
ont <- get_ontology("hp")
dat2 <- prune_ancestors(dat,id_col="hpo_id",ont=ont)
}
