% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0docs.R, R/query_oard.R
\name{query_}
\alias{query_}
\alias{query_oard}
\title{Query functions}
\source{
\href{https://smart-api.info/ui/25d608913f0e16a87dd29dabe3f52f78}{
OARD API docs}

\href{https://www.cell.com/ajhg/fulltext/S0002-9297(22)00319-6}{
Publication}
}
\usage{
query_oard(
  ids,
  concept_prefix = "q=",
  domain = "https://rare.cohd.io/api/",
  dataset_id = NULL,
  endpoint = "vocabulary/findConceptByAny",
  batch_size = 100
)
}
\arguments{
\item{ids}{IDs to query.}

\item{concept_prefix}{Concept prefix.}

\item{domain}{OARD domain URL.}

\item{dataset_id}{Dataset ID.}

\item{endpoint}{OARD API endpoint to use. 
See the \href{https://smart-api.info/ui/25d608913f0e16a87dd29dabe3f52f78}{
OARD API docs} for details.}

\item{batch_size}{Number of IDs to query at once.}
}
\value{
Queried data.
}
\description{
Functions to query specific subset of data via dedicated APIs.
}
\section{Functions}{
\itemize{
\item \code{query_oard()}: query_

Query the Open Real-world-based Annotation for Rare Disease (OARD).
See \href{https://rare.cohd.io/}{here} for details.

}}
\concept{query_}
