% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0docs.R, R/example_dat.R, R/get_data.R,
%   R/messager.R, R/report_filter.R, R/report_missing.R, R/stopper.R
\name{utils_}
\alias{utils_}
\alias{example_dat}
\alias{get_data}
\alias{messager}
\alias{report_filter}
\alias{report_missing}
\alias{stopper}
\title{Utility functions}
\usage{
example_dat(
  types = c("omim_id", "orphanet_id", "decipher_id", "hpo_id", "gene"),
  rm_types = NULL
)

get_data(
  file,
  tag = "latest",
  repo = "neurogenomics/KGExplorer",
  save_dir = cache_dir(),
  add_version = FALSE,
  overwrite = TRUE
)

messager(..., v = Sys.getenv("VERBOSE") != "FALSE", parallel = FALSE)

report_filter(g1, g2, cols, suffix = "remain after filtering.")

report_missing(dat, id_col, report_col)

stopper(..., v = TRUE)
}
\arguments{
\item{types}{Types of IDs to include in the data.}

\item{file}{Can be one of the following:
\itemize{
\item{"ontology"}{Creates an \link[simona]{ontology_DAG} R object by
 importing the OBO file directly from the official
 \href{https://github.com/obophenotype/upheno}{uPheno GitHub repository}.}
\item{"bestmatches"}{Returns a merged table with the best matches between
 human and non-human homologous phenotypes (from multiple species).
 Distributed by the official
 \href{https://github.com/obophenotype/upheno/tree/master/mappings}{
 uPheno GitHub repository}.}
\item{"upheno_mapping"}{Return a merged table with matches between human
and non-human homologous phenotypes (from multiple species).
Distributed by the
 \href{https://data.monarchinitiative.org/upheno2/current/upheno-release/all/index.html}{
 Monarch Initiative server}.}
}}

\item{tag}{tag for the GitHub release to which this data should be attached.}

\item{repo}{Repository name in format "owner/repo". Defaults to \code{guess_repo()}.}

\item{save_dir}{Directory to save a file to.}

\item{add_version}{Add the release version
to the returned object's \link[base]{attributes}}

\item{overwrite}{Should any local files of the same name be overwritten?
default \code{TRUE}.}

\item{...}{Additional arguments passed to plot-specific functions.}

\item{v}{Whether to print messages or not.}

\item{parallel}{Whether to enable message print when wrapped 
in parallelised functions.}
}
\value{
Merged data.

\link[data.table]{data.table}

Path to downloaded file or the object itself (when ".rds" format).

Null

Null
}
\description{
Functions to support other functions.
}
\section{Functions}{
\itemize{
\item \code{example_dat()}: utils_
Example data

Construct an example \link[data.table]{data.table} containing diseases, 
phenotype, and/or genes.

\item \code{get_data()}: utils_
Get remote data

Download remotely stored data via \link[piggyback]{pb_download}.

\item \code{messager()}: utils_
Print messages 

Conditionally print messages.
 Allows developers to easily control verbosity of functions, 
 and meet Bioconductor requirements that dictate the message 
 must first be stored to a variable before passing to \link[base]{message}.

\item \code{report_filter()}: utils_

\item \code{report_missing()}: utils_

\item \code{stopper()}: utils_
Stop messages

Conditionally print stop messages.
Allows developers to easily control verbosity of functions, 
 and meet Bioconductor requirements that dictate the stop message 
 must first be stored to a variable before passing to \link[base]{stop}.

}}
\examples{
dat <- example_dat()
}
\concept{utils_}
\keyword{internal}
